pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "magenta_rt"
version = "0.3.0"
authors = [
  { name="Google DeepMind" },
]
description = "Magenta RT: Real-time music generation"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
  "aiohttp",
  "absl-py",
  "chex",
  "clu",
  "flax",
  "gin-config",
  "flaxformer @ git+https://github.com/google/flaxformer@399ea3a",
  "google-cloud-storage",
  "jax",
  "jaxlib",
  "librosa",
  "numpy",
  "resampy",
  "requests",
  "sentencepiece",
  "seqio",
  "soundfile",
  "tf2jax",
  "tqdm",
  "typing-extensions",
]

[project.optional-dependencies]
test = [
  "pytest",
]
gpu = [
  "jax[cuda12]",
  "t5x[gpu]",
]
tpu = [
  "jax[tpu]==0.8.1",
  "t5x[tpu]",
]

[project.urls]
Homepage = "https://github.com/magenta/magenta-realtime"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"magenta_rt" = [
  "depthformer/configs/*.gin",
  "colab/static/js/*.js",
  "colab/static/html/*.html",
  ]
